master-service/requirements.txt
asgiref
Django
djangorestframework
mysqlclient
psycopg2-binary
python-decouple
setuptools
sqlparse
wheel
